feat: cleanup useTradeNetworkFeeCryptoBaseUnit
